fly away kit problem
this is a proper implementaiton with additional dimension that keeps the indicator for the iteration number. there are as many iterations as cargo types of course, and when we are in iteration i, it can happen that the optimal values of x_k's for k = 1:i are different from those 1:i values of x_k's as we are in iteration i+1.